test(opencode): split adapter behavior suites
